stash:

+++ src/Day4__SecureContainer.java
@@ -3,19 +3,19 @@ import interfaces.IntInputs;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 public class Day4__SecureContainer implements Day, IntInputs {
     @Override
     public int part_one() {
-        final int[] range = get_int_inputs("-");
-        return (int) IntStream.rangeClosed(range[0], range[1])
+        final long[] range = get_int_inputs("-");
+        return (int) LongStream.rangeClosed(range[0], range[1])
                 .filter(this::does_not_decrease)
                 .filter(this::has_same_adjacent_number)
                 .count();
     }
 
-    public boolean does_not_decrease(int i) {
+    public boolean does_not_decrease(long i) {
         String s = ("" + i);
         for (int j = 1; j < s.length(); j++) {
             if (s.charAt(j) < s.charAt(j - 1))
@@ -24,12 +24,12 @@ public class Day4__SecureContainer implements Day, IntInputs {
         return true;
     }
 
-    public boolean has_same_adjacent_number(int i) {
-        return !get_adjacent_lengths(i).isEmpty();
+    public boolean has_same_adjacent_number(long i) {
+        return !get_adjacent_lengths((int) i).isEmpty();
     }
 
-    public boolean has_isolated_pair(int i) {
-        return get_adjacent_lengths(i).contains(2);
+    public boolean has_isolated_pair(long i) {
+        return get_adjacent_lengths((int) i).contains(2);
     }
 
     private List<Integer> get_adjacent_lengths(int i) {
@@ -51,8 +51,8 @@ public class Day4__SecureContainer implements Day, IntInputs {
 
     @Override
     public int part_two() {
-        final int[] range = get_int_inputs("-");
-        return (int) IntStream.rangeClosed(range[0], range[1])
+        final long[] range = get_int_inputs("-");
+        return (int) LongStream.rangeClosed(range[0], range[1])
                 .filter(this::does_not_decrease)
                 .filter(this::has_isolated_pair)
                 .count();
